UF split
