+++ src/profile/v100/entity/account/mod.rs
@@ -1,7 +1,9 @@
 mod account;
 mod appearance_id;
+mod appearance_id_uniffi_fn;
 mod on_ledger_settings;
 
 pub use account::*;
 pub use appearance_id::*;
+pub use appearance_id_uniffi_fn::*;
 pub use on_ledger_settings::*;
